% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fantasy}
\alias{fantasy}
\title{Fantasy Scores From 2018 fantasy league}
\format{A data_frame with 132 rows and 3 variables
\describe{
  \item{team}{team name}
  \item{week}{week of season}
  \item{scores}{team score}
}}
\usage{
fantasy
}
\description{
A dataset containing scores, week and team name from fantasy league
}
\keyword{datasets}
