% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfw.R
\name{gen_team_dat}
\alias{gen_team_dat}
\title{Generate team data by week for testing}
\usage{
gen_team_dat(labels = paste0("team", 1:10), weeks = 10,
  distribution = "norm", decimal = F)
}
\arguments{
\item{labels}{vector of team labels}

\item{weeks}{total number of weeks}

\item{distribution}{list with elements, dist = c("norm"), mean, sd}

\item{decimal}{indicator for decimal scoring}
}
\description{
Generate team data by week for testing
}
