% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfw.R
\name{gen_team_wins}
\alias{gen_team_wins}
\title{Generate actual wins by week for testing}
\usage{
gen_team_wins(team_labels = paste0("team", 1:10), weeks = 10)
}
\arguments{
\item{team_labels}{vector of team labels}

\item{weeks}{total number of weeks}
}
\value{
data_frame with team and wins
}
\description{
Generate actual wins by week for testing
}
