% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfw.R
\name{plot_cumsfw}
\alias{plot_cumsfw}
\title{Plots cumulative SFW}
\usage{
plot_cumsfw(sfw_obj)
}
\arguments{
\item{sfw_obj}{sfw data_frame object from sfw()}
}
\value{
plot of SFW across week
}
\description{
Generates Longitudinal SFW Plot
}
