% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfw.R
\name{plot_scatsfw}
\alias{plot_scatsfw}
\title{Plots SFW vs Wins}
\usage{
plot_scatsfw(sfw_cum)
}
\arguments{
\item{sfw_cum}{sfw data_frame object from sfw(), cumulative object}
}
\value{
plot of SFW vs Wins
}
\description{
Generates Scatterplot of SFW vs Wins
}
