% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfw.R
\name{plot_sfw_wins}
\alias{plot_sfw_wins}
\title{Plots Difference SFW and Actual Wins}
\usage{
plot_sfw_wins(sfw_cum)
}
\arguments{
\item{sfw_obj}{sfw data_frame object from sfw(), cumulative}
}
\value{
plot of SFW vs Wins
}
\description{
Generates Barplots of SFW & Wins
}
