% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfw.R
\name{plot_windiff}
\alias{plot_windiff}
\title{Plots Difference between SFW and Wins - Barplots ... Luck}
\usage{
plot_windiff(sfw_cum)
}
\arguments{
\item{sfw_obj}{sfw data_frame object from sfw(), cumulative}
}
\value{
plot of SFW vs Wins
}
\description{
Generates Barplots of SFW vs Wins
}
