% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfw.R
\name{sfw}
\alias{sfw}
\title{Calculate Schedule Free Wins (SFW)}
\usage{
sfw(team_dat)
}
\arguments{
\item{team_dat}{tibble of teams scores by week, columns must include week, scores
(may be decimal), team labels, wins}
}
\value{
data_frame with team's SFW by week & cumulative. May include plots
(may add separate functions), bar plot of each teams SFW vs actual wins and scatterplot of
SFW vs actual wins
}
\description{
Schedule Free Wins Description
}
